All changes I have some clue for up to 1.91.9b
I have no clue how to fix the font and kerning stuffs

+++ Dalamud/Interface/Textures/Internal/TextureManager.BlameTracker.cs
@@ -2,7 +2,7 @@ using System.Collections.Generic;
 using System.Runtime.CompilerServices;
 using System.Runtime.InteropServices;
 
-using Dalamud.Bindings.ImGui;
+using Hexa.NET.ImGui;
 using Dalamud.Interface.Textures.TextureWraps;
 using Dalamud.Plugin.Internal.Types;
 using Dalamud.Plugin.Services;
@@ -261,7 +261,7 @@ internal sealed partial class TextureManager
 
                 this.srvDebugPreviewExpiryTick = Environment.TickCount64 + 1000;
                 if (!this.srvDebugPreview.IsEmpty())
-                    return new ImTextureID(this.srvDebugPreview.Get());
+                    return new ImTextureID((nint)this.srvDebugPreview.Get());
                 var srvDesc = new D3D11_SHADER_RESOURCE_VIEW_DESC(
                     this.tex2D,
                     D3D_SRV_DIMENSION.D3D11_SRV_DIMENSION_TEXTURE2D);
@@ -275,7 +275,7 @@ internal sealed partial class TextureManager
                     return Service<DalamudAssetManager>.Get().Empty4X4.Handle;
 
                 srv.Swap(ref this.srvDebugPreview);
-                return new ImTextureID(this.srvDebugPreview.Get());
+                return new ImTextureID((nint)this.srvDebugPreview.Get());
             }
         }
